fptools/libraries/HaXml: changeset 2003-05-13 13:05:33 by malcolm
Add tentative support for the ANY content spec. Obviously we don't
have a universal type in Haskell, so for the moment ANY converts to a
dummy (empty) type ANYContent, which throws away everything contained
within this element. This might be OK as an approximation, because an
ANY type is often declared as a way of including some random other XML
document, of which the processor need have no knowledge.