fptools/libraries/HaXml: changeset 2004-03-25 14:48:05 by malcolm
When encountering a parse failure, rather than terminate the whole
program with an explicit call to 'error', pass the error message up
through an Either type to the caller. The old API of the parser
combinators is still available - the new versions are primed, e.g.
papply', xmlParse', etc. Based on a suggestion (and some code) by
Graham Klyne.