more entityvalue fixes for PERefs
It turns out that an EntityValue /does/ need to be re-lexed for
the possible presence of PERefs.  However, this is a special case,
because the PERef must /not/ be separated from surrounding text by
logical whitespace, i.e. text+PERef can form a single lexical token
/after/ expansion of the PERef.  This was a previously unnoticed
change in the Second Edition of the XML spec.