big fixup for bad CVS->darcs conversion
The patch from version 1.13 to 1.14 of HaXml was incomplete, due to a
bad conversion between the CVS and darcs repositories.  I think this
large extra patch should fix it all up again.  Ideally, I'd like to
unrecord this and the previous patch, then re-record based on the CVS
logs to get a more fine-grained set of patches.  Until I work out how
to do that...