add "xmlParseWith" to the API, and export individual parsers (e.g. element)
Following a suggestion of Dmitry Astapov, for support of the XMPP
network protocol (RFC 3920), it is useful to permit partial parsing of
an input stream.  In the XMPP protocol, both client and server send
opening "<stream ...>" tags at the beginning of session, then send
lots of "<message>....</message>" packets during session, and close
session by sending "</stream>".  This is not the same as the SAX model,
and was previously not supported by HaXml.