In Xtract patterns, downcase HTML only for tags+attrnames, not attrvals.
Although XML is case-sensitive, HTML is case-insensitive regarding
tagnames and attribute names.  But of course the attribute *values* are
still case sensitive, so a blanket (map toLower) on pattern queries is
wrong.  Here is an example of a document and query which previously
failed:

   <a class="Foo">

   Xtract "//a[@class='Foo']"

The solution is to fix the lexer to do downcasing (or more generally, any
transformation) on the appropriate tokens.