Change meaning of "//" in Xtract query from deep to multi.
When a query contains a predicate, a tag-match for which the predicate
failed, caused searching to ignore the subtree.  This is due to the
semantics of the "deep" combinator.  "deepest" would look at the subtree
first, but this turns out to give surprising results for certain other
predicate queries.  Using "multi" ensures that both the local root _and_
its subtree are searched.  (However, one should be aware that the same
subtree might now occur in the results more than once, if the predicate
matches at both an inner and an outer level of nesting.)