Add a -n cmd-line option to Xtract.
With -n, the output converts entities like &lt; to their textual
equivalents e.g. <.  This applies to _all_ nested textual content within
elements (and in attributes too), making the output not strictly
XML-conformant, but more human-readable.

As a consequence of the way this is implemented, it also fixes an output
problem that used to occur with adjacent text and entities, where they
would appear on separate lines, e.g.
   don
   &apos;
   t
Now, all text is 'unescape'd internally during the query, and only
'escape'd again at the output stage (if -n is _not_ present).  The
net result is that the previous example now looks more correct:
   don&apos;t