Bugfix in DtdToHaskell: avoid stack overflow.
As reported by Bertram Felgenhauer, the following DTD
    <!ELEMENT dummy EMPTY>
    <!ELEMENT broken (#PCDATA | dummy)* >
when passed through DtdToHaskell, generated a parser that crashes on
    readXml "<broken>abc</broken>" :: Either String Broken
instead of giving the correct answer
    Just (Broken [Broken_Str "abc"])
The fix is to be more careful about generating the parsing expression
    (text `onFail` return "")
It is wrong to return the empty string inside a 'many' combinator,
because it simply loops forever.  But it is essential to return the empty
string if there is no enclosing zero-or-more combinator.