XsdToHaskell: resolve qualified names in attribute values correctly.
Previously, we just treated them as unqualified.  But we can resolve
them when parsing Schema, because we know which attrvals should be QNames,
and we have the global Namespace declarations too.  The one tricky bit is
to ensure that if the qualified name belongs to the target namespace for this
document, we must ensure it becomes unqualified (because of restrictions
in Haskell).