Correct logic-changing indentation in xquery
This code was changed as part of some hlint-suggested fixes, but it was
accidentally changed too much - the `onFail` was de-indented until it
also wrapped the parent `do` block, which changed the behaviour of the
code.

This changes it back by re-indenting the `onFail` so it only wraps the
innermost `do`.